% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantIso_model.r
\name{calculate_Tleaf}
\alias{calculate_Tleaf}
\title{get leaf temperature for
Interface with the tealeaves package in order to calculate leaf temperaure and boundary conductance.
Wrapper for tealeaves_getValues}
\usage{
calculate_Tleaf(isoplantspar)
}
\arguments{
\item{isoplantspar}{data.frame of randomized model parameters for the siensitvity analysis}
}
\value{
input datafrme with updated Tleaf and boundary resistance values
}
\description{
get leaf temperature for
Interface with the tealeaves package in order to calculate leaf temperaure and boundary conductance.
Wrapper for tealeaves_getValues
}
