% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_helper.r
\name{D2delta}
\alias{D2delta}
\title{Isotopic Delta conversion D to d}
\usage{
D2delta(smpl, ref)
}
\arguments{
\item{smpl}{D value fo sample}

\item{ref}{d value fo reference}
}
\value{
d notation of sample over reference
}
\description{
Isotopic Delta conversion D to d
}
