% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_helper.r
\name{delta2D}
\alias{delta2D}
\title{Isotopic Delta conversion small d to D}
\usage{
delta2D(smpl, ref)
}
\arguments{
\item{smpl}{d value fo sample}

\item{ref}{d value fo reference}
}
\value{
D notation of sample over reference
}
\description{
Isotopic Delta conversion small d to D
}
