% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_fitting.r
\name{fit_plant18O}
\alias{fit_plant18O}
\title{Calculate RMSE for parameter fitting with isoplants}
\usage{
fit_plant18O(
  fitparnames,
  data,
  obsvalue,
  modvalue = "d18O_c",
  method = "L-BFGS-B",
  pd = NULL
)
}
\arguments{
\item{fitparnames}{vecort of value names to be fitted}

\item{data}{data.frame of plant18O model parameters (except the ones to be fitted)
including a column of observation data}

\item{obsvalue}{name of the column in fixpar datafrme contoing the observation data}

\item{modvalue}{name of the output variable to be tested againtst}

\item{method}{optimization methods. Supported methods are all methods in \code{optim} or 'GenSA' , which is more likely to find the glabal optimum in complex parameter space}

\item{pd}{optional parameter definition data.frame (pd <- get_parameter_definition()) with custom boundaries}
}
\value{
output from optimization routine including best fitting parameters
}
\description{
Calculate RMSE for parameter fitting with isoplants
}
