% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_fitting.r
\name{fit_plant2H}
\alias{fit_plant2H}
\title{Estimate parameters of the hydrogen isotope model by optimization procedures}
\usage{
fit_plant2H(
  fitparnames,
  data,
  obsvalue,
  modvalue = "d2H_c",
  method = "L-BFGS-B",
  pd = NULL,
  leaftemperature = FALSE,
  verbose = TRUE
)
}
\description{
A wrapper for \code{fit_plantIso()} to fit the oxygen isotope model. See \code{?fit_plantIso()} for detailed description of arguments.
}
