% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_fitting.r
\name{fit_plantIso}
\alias{fit_plantIso}
\title{Estimate parameters of the isotope model by optimization prodecures}
\usage{
fit_plantIso(
  fitparnames,
  data,
  obsvalue,
  modvalue = "d18O_c",
  element = "O",
  method = "L-BFGS-B",
  pd = NULL,
  leaftemperature = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{fitparnames}{vecort of value names to be fitted}

\item{data}{data.frame of plant18O model parameters (except the ones to be fitted)
including a column of observation data}

\item{obsvalue}{name of the column in fixpar datafrme contoing the observation data}

\item{modvalue}{name of the output variable to be tested againtst}

\item{element}{additional parameters, run for oxygen 'O' or hydrogen 'H'}

\item{method}{optimization methods. Supported methods are all methods in \code{optim} or 'GenSA' , which is more likely to find the glabal optimum in complex parameter space}

\item{pd}{optional parameter definition data.frame (pd <- get_parameter_definition()) with custom boundaries}

\item{leaftemperature}{boolean. Calculate leaf temperature using the tealeaves model. Parameters 'swrad', 'wind' and 'leafsize must be provided'}

\item{verbose}{boolean. Output some information while parameter fitting}
}
\value{
output from optimization routine including best fitting parameters
}
\description{
Estimate parameters of the isotope model by optimization prodecures
}
