% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantIso_model.r
\name{get_default_parameters}
\alias{get_default_parameters}
\title{Get a set of default environment and leaf parameters}
\usage{
get_default_parameters(
  n = 1,
  mode = "peclet",
  element = "O",
  tealeaves = FALSE
)
}
\arguments{
\item{n}{number of rows in the retuned data.frame}

\item{mode}{'peclet' (default) or 'twopool', specify the mixing model used}

\item{element}{get parameters for oxygen 'O' or hydrogen 'H'}

\item{tealeaves}{add default parameters for leaf temperature calculation with the tealeaves model}
}
\value{
a dataframe containing vaild parameter names,  ranges, descriptors and default values
}
\description{
Get a set of default environment and leaf parameters
}
