% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantIso_model.r
\name{get_equilibrium_fractionation}
\alias{get_equilibrium_fractionation}
\title{Equilibrium fractionation}
\usage{
get_equilibrium_fractionation(Tleaf, element = "O")
}
\arguments{
\item{Tleaf}{Leaf temperature in Celsius}

\item{element}{specify whether it should be calculated for oxygen 'O' of hyrdogen 'H'}
}
\value{
equilibrium_fractionation in permil
}
\description{
Temperature dependence of equilibrium fractinations see Majoube 1971.
implemented as described in review by Cernusak et al. 2016
}
\references{
Cernusak, L.A., Barbour, M.M., Arndt, S.K., Cheesman, A.W., English, N.B., Feild, T.S., Helliker, B.R., Holloway-Phillips, M.M., Holtum, J.A., Kahmen, A. and McInerney, F.A., 2016. Stable isotopes in leaf water of terrestrial plants. Plant, Cell & Environment, 39(5), pp.1087-1102.
}
