% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantIso_model.r
\name{get_ewc}
\alias{get_ewc}
\title{Equilibrium fractionation between carbonyl oxygen and medium water}
\usage{
get_ewc(Tleaf)
}
\arguments{
\item{Tleaf}{Leaf temperature}
}
\value{
ewc value in permil
}
\description{
calculate ewc after by Sternberg & Ellsworth (2011)
}
