% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantIso_model.r
\name{get_kinetic_fractionation}
\alias{get_kinetic_fractionation}
\title{Kinetic fractionation}
\usage{
get_kinetic_fractionation(rs, rb, element = "O")
}
\arguments{
\item{rs}{stomatal resistance}

\item{rb}{boundary layer resistance}

\item{element}{specify whether it sould be calculated for oxygen 'O' of hyrdogen 'H'}
}
\value{
kinetic_fractionation in permil
}
\description{
Resistance (stomatal/boundary layer) dependence of kinetic fractinations.
implemented as described in review by Cernusak et al. 2016:
ek <- (28 x rs + 19 x rb) / (rs + rb)
other suggest:
ek<-(32 x rs+22 x rb)/(rs+rb) #described in review by Barbour 2007
ek<-(32 x rs+21 x rb)/(rs+rb) #as described in paper by Song 2013
}
\references{
Cernusak, L.A., Barbour, M.M., Arndt, S.K., Cheesman, A.W., English, N.B., Feild, T.S., Helliker, B.R., Holloway-Phillips, M.M., Holtum, J.A., Kahmen, A. and McInerney, F.A., 2016. Stable isotopes in leaf water of terrestrial plants. Plant, Cell & Environment, 39(5), pp.1087-1102.
}
