% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant18O_model.r
\name{get_parameter_definition}
\alias{get_parameter_definition}
\title{Basic description of input parameters for the leaf18O-model
This data.frame object returned by this function dataframe contains basic information about model inpiut parameters and is used for the check_parameters function}
\usage{
get_parameter_definition()
}
\value{
a dataframe containing vaild parameter names,  ranges, descriptors and default values
}
\description{
Basic description of input parameters for the leaf18O-model
This data.frame object returned by this function dataframe contains basic information about model inpiut parameters and is used for the check_parameters function
}
\examples{

\dontrun{
parameters = get_parameter_definition ()
}
}
\keyword{cellulose,}
\keyword{craig-gordon,}
\keyword{isotope,}
\keyword{leaf}
\keyword{parameters}
\keyword{water,}
