% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantIso_model.r
\name{get_pathlength}
\alias{get_pathlength}
\title{Transpiration dependent scaled path length}
\usage{
get_pathlength(E)
}
\arguments{
\item{E}{Transpiration}
}
\value{
path length value in m
}
\description{
calculate path length after song et al. 2013
}
