% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantIso_model.r
\name{get_peclet_number}
\alias{get_peclet_number}
\title{Peclet number}
\usage{
get_peclet_number(Lm, E, D)
}
\arguments{
\item{Lm}{scaled path lenth, i m}

\item{E}{transpiration in mol m-2 s-1}

\item{D}{Diffusivity}
}
\value{
peclet number
}
\description{
Peclet number
}
