% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantIso_model.r
\name{get_pressure_at_elevation}
\alias{get_pressure_at_elevation}
\title{atmospheric pressure at elevation}
\usage{
get_pressure_at_elevation(elevation, Tair)
}
\arguments{
\item{elevation}{elevation in m}

\item{Tair}{Air temperature in C}
}
\value{
a dataframe containing vaild parameter names,  ranges, descriptors and default values
}
\description{
calculate mean pressure based on mean sea level pressure (1013.25 mb) and elevation
this approximation doesn't take changing atmospheric pressure into account, but these changes have a very minor impact on calculated
The main thing is to get the pressure changes due to elevation.
}
