% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantIso_model.r
\name{get_randomized_parameters}
\alias{get_randomized_parameters}
\title{sample parameters from specified distribution for leaf18O model}
\usage{
get_randomized_parameters(n, parranges)
}
\arguments{
\item{n}{number of parameter sets}

\item{parranges}{data.frame (name,pdist,pdm,pdv)}
}
\value{
data.frame containing parameters sampled from the defined distrubution
}
\description{
sample parameters from specified distribution for leaf18O model
}
\examples{
\dontrun{
# RUN 1000 more time on a randomly varied parameters
n<-10000
rndpar<-get_randomized_parameters(n,parranges)
check_parameters(rndpar)[[2]]
modelout<-plant18O_model(rndpar,output='d18O_pt')
quantile(modelout,c(0.025,0.25,0.5,0.75,0.975))
mean(modelout[!is.infinite(modelout)])
}

}
