% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_fitting.r
\name{get_run_parameters}
\alias{get_run_parameters}
\title{adds fitted parameter to constant data}
\usage{
get_run_parameters(constpar, fitpar)
}
\arguments{
\item{constpar}{data.frame containing static values used to run the model}

\item{fitpar}{named vector containing fitted values used to run the model}
}
\value{
data.frame containing all values used to run the model
}
\description{
adds fitted parameter to constant data
}
