% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantIso_model.r
\name{get_saturation_vapor_pressure}
\alias{get_saturation_vapor_pressure}
\title{Saturation Vapor pressure}
\usage{
get_saturation_vapor_pressure(T)
}
\arguments{
\item{T}{Temperature in Celsius}
}
\value{
saturation Vapor pressure in kPa
\deqn{svp = h_\mathrm{vap} g_\mathrm{tw} d_\mathrm{wv}}{L = h_vap g_tw d_wv}
}
\description{
Saturation Vapor pressure
}
