% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant18O_model.r
\name{get_stomatal_conductance}
\alias{get_stomatal_conductance}
\title{Stomatal conductance}
\usage{
get_stomatal_conductance(gsref, vpd)
}
\arguments{
\item{gsref}{stomatal conductance reference, measured at 1 kPa VPD}

\item{vpd}{vapour pressure deficit}
}
\value{
stomatal_conductance in mol m-2 s-1
}
\description{
Stomatal conductance as calculated by the formula of Oren 1999
}
