% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant18O_model.r
\name{get_transpiration}
\alias{get_transpiration}
\title{Transpiration}
\usage{
get_transpiration(VPD, P, rs, rb)
}
\arguments{
\item{VPD}{Vapor pressure deficit}

\item{P}{Atmospheric pressure}

\item{rs}{stomatal resistance}

\item{rb}{boundary layer resistance}
}
\value{
Leaf transpiration in mol m^-2 s^-2
}
\description{
Transpiration
}
