% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisoAI_APIclient.r
\name{pisoai_get_data}
\alias{pisoai_get_data}
\title{Get multi site data through the Piso.AI API}
\usage{
pisoai_get_data(
  location,
  years = c(1950, 2019),
  months = c(1, 12),
  storelocal = NA
)
}
\arguments{
\item{location}{data.frame object with colums 'site','latitude','longitude'. Colums 'elevation' and 'year' are optional}

\item{years}{single year or vector with range of years (default c(1950,2019))}

\item{months}{single month or vector with range of months (default c(1,12))}

\item{storelocal}{optional path to store downloaded data (default NA, files are not stored), use pisoai_readlocal() to read local data}
}
\value{
data.frame with PISO.AI prediction for precipitation isotopic composition
}
\description{
\href{https://isotope.bot.unibas.ch/PisoAI/}{Piso.AI} is tool for predicting monthly timeseries of oxygen and hydrogen isotope values of precipitation that uses a machine learning model trained on geographic and climate data.
}
\examples{

# Piso.AI data may be requested as
\dontrun{
 # Simple request for Piso.AI data of two locations and the full time span of Piso.AI
 location<-data.frame(site= c("TEST1","TEST2") ,latitude=c(45.7,46.5),longitude=c(7.6,8.0))
 pisoai_data <- pisoai_get_data(location)


 # Request for Piso.AI data of two locations and specific timespan
 location<-data.frame(site= c("TEST1","TEST2") ,latitude=c(45.7,46.5),longitude=c(7.6,8.0))
 pisoai_data <- pisoai_get_data(location, years=c(1990,2000),months=c(4,6))

 # Request for Piso.AI data of two locations individual years
 location<-data.frame(site= c("TEST1","TEST2"),
                      latitude=c(45.7,46.5),longitude=c(7.6,8.0),
                      years=c(1979,2010))
 pisoai_data <- pisoai_get_data(location)

 # Similar request, additionally storing data locally
 pisoai_data <- pisoai_get_data(location,storelocal='~/PisoAI_data')
 # read back local data
 pisoai_data <- pisoai_readlocal('~/PisoAI_data')
}
}
