% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisoAI_APIclient.r
\name{pisoai_request}
\alias{pisoai_request}
\title{simple Piso.AI API request}
\usage{
pisoai_request(
  site,
  latitude,
  longitude,
  elevation = NA,
  years = c(1950, 2019),
  months = c(1, 12),
  storelocal = NA
)
}
\arguments{
\item{site}{site name}

\item{latitude}{latitude in degrees}

\item{longitude}{latitude in degrees}

\item{elevation}{optional custom elevation (otherwise it will be extracted from DEM)}

\item{years}{single year or vector with range of years (default c(1950,2019))}

\item{months}{single month or vector with range of months (default c(1,12))}

\item{storelocal}{optional path to store downloaded data (default NA, files are not stored), use pisoai_readlocal() to read local data}
}
\value{
data.frame with PISO.AI prediction for precipitation isotopic composition
}
\description{
\href{https://isotope.bot.unibas.ch/PisoAI/}{Piso.AI} is tool for predicting monthly timeseries of oxygen and hydrogen isotope values of precipitation that uses a machine learning model trained on geographic and climate data.
}
\examples{

# single site Piso.AI data may be requested as
\dontrun{
 pisoai_reqest('BSL',45.7,7.6)
}
}
