% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant18O_model.r
\name{plant18O_model}
\alias{plant18O_model}
\title{(Péclet mofified) Craig-Gordon based model for plant tissue 18O enrichment
Calculates d18O of leaf water, cellulose and bulk tissue with different models based on the parameter provided}
\usage{
plant18O_model(par, addpar = NULL, output = "all", verbose = FALSE)
}
\arguments{
\item{par}{a named vector or data.frame of parameter values (see details below)}

\item{addpar}{additional parameters, that will my join internally with the main parameter list. To be used for fixed parameters for model fitting/sensitivity analysis}

\item{output}{vector of output selected output values or 'all' (default)}

\item{verbose}{print messages while running}
}
\value{
a dataframe of calculated values (see details below)
}
\description{
(Péclet mofified) Craig-Gordon based model for plant tissue 18O enrichment
Calculates d18O of leaf water, cellulose and bulk tissue with different models based on the parameter provided
}
\details{
\bold{Input parameters}\cr A data.frame with the following numeric columns:

\tabular{llll}{
\bold{Environment}\cr
\code{Tair} \tab Air Temperature (C)\cr
\code{RH} \tab Relative humidity (\%)\cr
\tab or\cr
\code{ea} \tab atmospheric vapor pressure  (kPa) \cr
\code{P} \tab barometric pressure (kPa) \cr
\tab or\cr
\code{elevation} \tab elevation (m) \tab calculate P(elevation)\cr
\code{d18O_sw} \tab d18O soil water (permille)\cr
\code{D18O_wv} \tab D18O of water vapor over soil water (permille) \tab optional, if not supplied water vapour is assubed to be in equilibrium with soil water\cr
\cr
\bold{Leaf}\cr
\code{DTleaf} \tab delta leaf temperature (K)\cr
\tab or\cr
\code{Tleaf} \tab absolute leaf temperature \tab C\cr
\code{rb} \tab boundary resistance (m^2 s mol^-1) \cr
\code{gs} \tab stomatal conductance (mol m^-2 s^-1)\cr
\tab or\cr
\code{gsref} \tab ref. stomatal conductance @ 1kPa VPD (mol m^-2 s^-1) \tab calculate gs as gs(vpd)\cr
\code{Lm} \tab Peclet-model scaled path length (m) \tab use Peclet-model\cr
\tab or\cr
\code{phi} \tab Two pool model mixing parameter \tab use Two pool-model\cr
\code{px} \tab Proportion of O exchanged during cellulose synthesis\cr
\code{pex} \tab Proportion of unenriched xylem water in developing cell\cr
\code{ewc} \tab Biosynthetic fracination during cellulose synthesis (assumed to be 27 permil)\cr
\code{ecp} \tab offset from cellulose to bulk leaf material\cr
\cr
\bold{additional leaf temperature model \code{tealeaves} parameter}\cr
\code{swrad} \tab shortwave radiation \tab (Wm^-2)\cr
\code{wind} \tab windspeed  \tab (ms^-1)\cr
\code{ls} \tab leaf size  \tab (m)\cr
}
}
\examples{

# The model returns the ....
\dontrun{
result = plant18O_model(par)
}
}
\keyword{cellulose,}
\keyword{craig-gordon,}
\keyword{isotope,}
\keyword{leaf}
\keyword{water}
