% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backsolve_model.r
\name{plant18O_solve_gs}
\alias{plant18O_solve_gs}
\title{Solves the PMCG model for gs}
\usage{
plant18O_solve_gs(par, d18Oc_colname = "d18O_c")
}
\arguments{
\item{par}{a named vector or data.frame of model parameters (see description oablve) and (observed) d18O_cellulose values}

\item{d18Oc_colname}{name of column in the input parameter vector/dataframe containg the d18O_cellupose values used for calculation  (default 'd18O_c')}
}
\value{
vector of calculated gs values
}
\description{
This function solves the PMCG model for gs, provided that the necessary parameters are supplied to the function
}
\details{
\bold{Input parameters}\cr A data.frame with the following numeric columns:

\tabular{llll}{
\bold{Environment}\cr
\code{Tair} \tab Air Temperature (C)\cr
\code{RH} \tab Relative humidity (\%)\cr
\code{P} \tab barometric pressure (kPa) \cr
\code{d18O_sw} \tab d18O soil water (permille)\cr
\cr
\bold{Leaf}\cr
\code{Tleaf} \tab absolute leaf temperature \tab C\cr
\code{rb} \tab boundary resistance (m^2 s mol^-1) \cr
\code{gs} \tab stomatal conductance (mol m^-2 s^-1)\cr
\code{Lm} \tab Peclet-model scaled path length (m) \tab use Peclet-model\cr
\code{px} \tab Proportion of O exchanged during cellulose synthesis\cr
\code{pex} \tab Proportion of unenriched xylem water in developing cell\cr
\code{ewc} \tab Biosynthetic fracination during cellulose synthesis (assumed to be 27 permil)\cr
\cr
\bold{Isotope signal}\cr
\code{d18O_cellulose} \tab Measured d18O_cellulose values, columnname may be specified as function parameter\cr
}
}
\examples{

\dontrun{
#Test
par <- get_default_parameters(n=3)
par$gs <- seq(0.3,0.5,0.1)                             # set test gs values
par$d18O_c <- plant18O_model(par,output = 'd18O_c')    # calculate d18O with test gs values
plant18O_solve_gs(par)                                 # backsolve model for gs
}


}
