% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant18O_model.r
\name{plantIso_model}
\alias{plantIso_model}
\title{(Péclet mofified) Craig-Gordon and based model for isotopic enrichment in plant tissue (18O and 2H)}
\usage{
plantIso_model(
  par,
  addpar = NULL,
  element = "O",
  output = "all",
  verbose = FALSE
)
}
\arguments{
\item{par}{a named vector or data.frame of parameter values (see details below)}

\item{addpar}{additional parameters, that will my join internally with the main parameter list. To be used for fixed parameters for model fitting/sensitivity analysis}

\item{element}{additional parameters, run for oxygen 'O' or hydrogen 'H'}

\item{output}{vector of output selected output values or 'all' (default)}

\item{verbose}{print messages while running}
}
\value{
a dataframe of calculated values (see details below)
}
\description{
(Péclet mofified) Craig-Gordon and based model for isotopic enrichment in plant tissue (18O and 2H)
}
\details{
Main code for the model, for detailed input parameter descriptions see ?plant18O_model() or ?plant2H_model() functions
}
\keyword{cellulose,}
\keyword{craig-gordon,}
\keyword{isotope,}
\keyword{leaf}
\keyword{water}
