% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_fitting.r
\name{plantIso_rmse}
\alias{plantIso_rmse}
\title{Calculate RMSE for parameter fitting with isoplants}
\usage{
plantIso_rmse(
  fitpar,
  fitparnames,
  constpar,
  obsvalue,
  modvalue = "d18O_c",
  element = "O",
  leaftemperature = FALSE
)
}
\arguments{
\item{fitpar}{data.frame containing fitted values used to run the model}

\item{fitparnames}{names of these values}

\item{constpar}{static parmeters that are reruired to run the model but are not part of the fitting process}

\item{obsvalue}{name of the column in fixpar datafrme contoing the observation data}

\item{modvalue}{name of the output variable to be tested againtst}

\item{element}{additional parameters, run for oxygen 'O' or hydrogen 'H'}

\item{leaftemperature}{name of the output variable to be tested againtst}
}
\value{
RMSE value
}
\description{
Calculate RMSE for parameter fitting with isoplants
}
