% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.r
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Plot the relative sensitivity of each factor along a gradient.
This function plots the output of scan_sensitivity()}
\usage{
plot_sensitivity(
  ds,
  xlab,
  ylab,
  main,
  fcolors = c("#d53e4f", "#fc8d59", "#fee08b", "#ffffbf", "#e6f598", "#99d594",
    "#3288bd"),
  parcol = 1,
  legend = FALSE
)
}
\arguments{
\item{ds}{one of the objects in the list created by scan_sensitivity()}

\item{xlab}{data.frame of randomized model parameters for the sensitivity analysis}

\item{ylab}{data.frame of randomized model parameters for the sensitivity analysis}

\item{main}{data.frame of randomized model parameters for the sensitivity analysis}

\item{fcolors}{colors for the different factors (default colors as defined in #http://colorbrewer2.org/#type=diverging&scheme=Spectral&n=6)}

\item{parcol}{index of the gradient color, will be hold out from the plot (default=1)}

\item{legend}{boolean. plot legend (default=FALSE)}
}
\description{
Plot the relative sensitivity of each factor along a gradient.
This function plots the output of scan_sensitivity()
}
\examples{
\dontrun{
plot_sensitivity(ds,scanpar,'normalized first-order indices','SA PMCG for D18O_lw',parcol = 1)
}
}
