% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.r
\name{scan_sensitivity}
\alias{scan_sensitivity}
\title{Sensitivity analysis by factor
Calculates sobol indices along scanning one parameter}
\usage{
scan_sensitivity(X1, X2, fixedPar, ts, scanpar, outpar = "D18O_lw")
}
\arguments{
\item{X1}{data.frame of randomized model parameters for the siensitvity analysis}

\item{X2}{data.frame of randomized model parameters for the siensitvity analysis}

\item{fixedPar}{one row data.frame with requiered parameters not to be varied in the sensitivity analysis}

\item{ts}{vector of values to be calculated for the gradient factor}

\item{scanpar}{character. Name of the Gradient factor}

\item{outpar}{name of the response variable (default "D18Olw")}
}
\value{
a list with two dataframes (1) the estimations of the Sobol' first-order indices (2) the estimations of the Sobol' total sensitivity indices.
}
\description{
Sensitivity analysis by factor
Calculates sobol indices along scanning one parameter
}
\examples{

\dontrun{
parameters = get_parameter_definition ()
}
}
