% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant18O_model.r
\name{set_parameters}
\alias{set_parameters}
\title{Set parameters by replacing or adding parameters to parameter table}
\usage{
set_parameters(newvalues, parameter_table)
}
\arguments{
\item{newvalues}{data.frame with new parameter values and columnnames machting a model parameter}

\item{parameter_table}{data.frame of parameters (same format as returned by get_default_parameters())}
}
\value{
updated parameter data.frame
}
\description{
if multiple values are assigned to d parameter table with only a single row, the nuber of rows will be extended to match the number of rows in the newparameter table.
}
