% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant18O_model.r
\name{tealeaves_getValues}
\alias{tealeaves_getValues}
\title{get leaf temperature for
Interface with the tealeaves package in order to calculate leaf temperaure and boundary conductance. Sets air temperauture,
barometric pressure, relative humidity and stomatal conductance from isoplants parameters}
\usage{
tealeaves_getValues(
  isoplantspar,
  tealeaves_enviro_par = NULL,
  tealeaves_leaf_par = NULL,
  tealeaves_constants = NULL
)
}
\arguments{
\item{isoplantspar}{data.frame of randomized model parameters for the siensitvity analysis}

\item{tealeaves_enviro_par}{(optional) customized tealeaves parameterset}

\item{tealeaves_leaf_par}{(optional) customized tealeaves parameterset}

\item{tealeaves_constants}{(optional) customized tealeaves parameterset}
}
\value{
vector containing Tleaf and rb value
}
\description{
get leaf temperature for
Interface with the tealeaves package in order to calculate leaf temperaure and boundary conductance. Sets air temperauture,
barometric pressure, relative humidity and stomatal conductance from isoplants parameters
}
