% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant18O_model.r
\name{apply_craig_gordon_model}
\alias{apply_craig_gordon_model}
\title{Craig-Gordon model for evaporative isotope enrichment}
\usage{
apply_craig_gordon_model(eq, ek, Dwv, ea, ei)
}
\arguments{
\item{eq}{equilibrium fractination (permil)}

\item{ek}{kinetic fractination (permil)}

\item{Dwv}{Water vapor isotope signal over source water (oxygen or hydrogen, depending on application)}

\item{ea}{atmospheric water pressure}

\item{ei}{leaf internal water pressure (assumed to be saturated)}
}
\value{
D18O at evaporative front over source water
}
\description{
Models the heavy isotope enchichment due to evaporation
}
