% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant18O_model.r
\name{apply_peclet_1D}
\alias{apply_peclet_1D}
\title{1D Peclet mixing}
\usage{
apply_peclet_1D(D18O_e, pn)
}
\arguments{
\item{D18O_e}{d18O value at evaporative front}

\item{pn}{peclet number}
}
\value{
D18O of total leaf water from d18O value at evaporative front
}
\description{
reduces the d18O at evaporative front based on the peclet value (Farquhar & Lloyd 1993)
}
