% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant18O_model.r
\name{apply_twopool_mixing}
\alias{apply_twopool_mixing}
\title{Two pool mixing
two pool model phi is the proportion of unenriched water}
\usage{
apply_twopool_mixing(phi, D18O_e)
}
\arguments{
\item{phi}{mixing coefficient}

\item{D18O_e}{D18O of water at the evaporative front}
}
\value{
D18O of total leaf water from d18O value at evaporative front
}
\description{
Two pool mixing
two pool model phi is the proportion of unenriched water
}
