% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant18O_model.r
\name{get_cellulose18O}
\alias{get_cellulose18O}
\title{Leaf water to cellulose 18O}
\usage{
get_cellulose18O(pex, px, D18O_lw, ewc = 27)
}
\arguments{
\item{pex}{exchange of leaf water signal in carbohydrates with local water before incorporation into the cellulose polymer}

\item{px}{fraction of unenriched water in this water pool used for cellulose synthesis}

\item{D18O_lw}{D180 of leaf water}

\item{ewc}{Equilibrium fractionation between carbonyl oxygen and medium water (default 27 permil)}
}
\value{
Returns Oxygen isotope composition of cellulose (permil)
}
\description{
Leaf water to cellulose 18O
}
