% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant18O_model.r
\name{get_H2O_diffusivity}
\alias{get_H2O_diffusivity}
\title{Diffusivity h2(18O)}
\usage{
get_H2O_diffusivity(T)
}
\arguments{
\item{T}{Temperature in Celsius}
}
\value{
diffusivity of heavy isotopologue  in water (H2 18O)
}
\description{
Temperature (C) dependence of diffusivity of heavy isotopologue  in water (H2
18O) (Cuntz et al. 2007)
as described in review by Cernusak et al. 2016
}
